move data transformation to handler
